requirements.txt
requests>=2.31.0
PyYAML>=6.0
flask>=3.0.0
pytest>=8.0.0
